+++ backend/canisters/local_user_index/api/src/queries/access_token_v2.rs
@@ -1,11 +1,10 @@
-use candid::CandidType;
 use oc_error_codes::OCError;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
 use types::{BotActionScope, BotCommandArg, BotCommandMeta, Chat, UserId, VideoCallType};
 
 #[ts_export(local_user_index, access_token_v2)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub enum Args {
     StartVideoCall(StartVideoCallArgs),
     JoinVideoCall(JoinVideoCallArgs),
@@ -14,7 +13,7 @@ pub enum Args {
 }
 
 #[ts_export(local_user_index, access_token_v2)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub enum Response {
     Success(String),
     NotAuthorized,
@@ -23,26 +22,26 @@ pub enum Response {
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct StartVideoCallArgs {
     pub call_type: VideoCallType,
     pub chat: Chat,
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct JoinVideoCallArgs {
     pub chat: Chat,
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct MarkVideoCallAsEndedArgs {
     pub chat: Chat,
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct BotActionByCommandArgs {
     pub bot_id: UserId,
     pub command: BotCommandInitial,
@@ -50,7 +49,7 @@ pub struct BotActionByCommandArgs {
 }
 
 #[ts_export]
-#[derive(CandidType, Serialize, Deserialize, Debug, Clone, PartialEq)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct BotCommandInitial {
     pub name: String,
     pub args: Vec<BotCommandArg>,